backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
anthropic==0.39.0
python-dotenv==1.0.1
pydantic==2.9.2
